Refactor project structure to add executors
