+++ graphql_subscriptions/__init__.py
@@ -3,4 +3,3 @@ from .subscription_manager import RedisPubsub, SubscriptionManager
 from .subscription_transport_ws import SubscriptionServer
 
 __all__ = ['RedisPubsub', 'SubscriptionManager', 'SubscriptionServer']
-
