+++ graphql_subscriptions/subscription_transport_ws/__init__.py
@@ -1,0 +1,6 @@
+from .base import BaseSubscriptionServer
+from ..executors.gevent import GeventMixin
+
+
+class SubscriptionServer(BaseSubscriptionServer, GeventMixin):
+    pass
