+++ graphql_subscriptions/subscription_transport_ws/protocols.py
@@ -1,0 +1,10 @@
+SUBSCRIPTION_FAIL = 'subscription_fail'
+SUBSCRIPTION_END = 'subscription_end'
+SUBSCRIPTION_DATA = 'subscription_data'
+SUBSCRIPTION_START = 'subscription_start'
+SUBSCRIPTION_SUCCESS = 'subscription_success'
+KEEPALIVE = 'keepalive'
+INIT = 'init'
+INIT_SUCCESS = 'init_success'
+INIT_FAIL = 'init_fail'
+GRAPHQL_SUBSCRIPTIONS = 'graphql-subscriptions'
